+++ CHANGELOG.md
@@ -2,6 +2,17 @@ This is a list of new fixes and other changes made to SOUP during its continued 
 
 SOUP version numbering generally follows the _major.minor.patch_ scheme.  Odd minor versions (1.1.x, 1.3.x, 1.5.x, etc.) are reserved for unstable development versions ("devel" branch on [GitHub][github]), while even ones denote stable releases ("master" branch on [GitHub][github]).  The change log below thus only includes even version numbers, as changes between development versions are too numerous to list.
 
+1.35 (development)
+====
+
+**New fixes in 1.35.x:**
+
+* [mso300679: Speaking out instead of suffering silently](http://meta.stackoverflow.com/q/300679)
+
+**Other changes:**
+
+* The [graphicdesign2415]((http://meta.graphicdesign.stackexchange.com/q/2415) CSS fix (for links in post warnings) is now applied globally.
+
 1.34 (31 Aug 2015)
 ====